requirements.txt
enum34==1.0.4
jira==1.0.3
oauthlib==1.0.3
PyGithub==1.25.2
requests==2.8.1
requests-oauthlib==0.5.0
requests-toolbelt==0.4.0
simplejson==3.8.1
six==1.10.0
tlslite==0.4.9
wheel==0.26.0
